IDEN-446 Add Solidity linder and autofix contracts

+++ contracts/examples/ERC20Verifier.sol
@@ -13,7 +13,8 @@ contract ERC20Verifier is ERC20, ZKPVerifier {
     mapping(uint256 => address) public idToAddress;
     mapping(address => uint256) public addressToId;
 
-    uint256 public TOKEN_AMOUNT_FOR_AIRDROP_PER_ID = 5 * 10**uint(decimals());
+    uint256 public TOKEN_AMOUNT_FOR_AIRDROP_PER_ID =
+        5 * 10**uint256(decimals());
 
     constructor(string memory name_, string memory symbol_)
         ERC20(name_, symbol_)
