+++ contracts/validators/CredentialAtomicQueryMTPValidator.sol
@@ -73,7 +73,7 @@ contract CredentialAtomicQueryMTPValidator is
             "wrong query operator has been used for proof generation"
         );
 
-         for (uint i = 0; i < query.value.length; i++) {
+        for (uint256 i = 0; i < query.value.length; i++) {
             require(
                 inputs[i + 10] == query.value[i],
                 "wrong comparison value has been used for proof generation"
