+++ contracts/validators/CredentialAtomicQuerySigValidator.sol
@@ -76,7 +76,7 @@ contract CredentialAtomicQuerySigValidator is
             "wrong query operator has been used for proof generation"
         );
 
-        for (uint i = 0; i < query.value.length; i++) {
+        for (uint256 i = 0; i < query.value.length; i++) {
             require(
                 inputs[i + 10] == query.value[i],
                 "wrong comparison value has been used for proof generation"
